compression refactor
